requirements.txt
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
coverage==5.5
cycler==0.10.0
Flask==2.0.1
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
nltk==3.6.2
numpy==1.20.3
packaging==21.0
pandas==1.2.4
Pillow==8.2.0
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.11.1
python-dateutil==2.8.1
python-dotenv==0.19.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
stanza==1.2.1
toml==0.10.2
torch==1.9.0
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
wordcloud==1.8.1
xlrd==2.0.1
blinker==1.4
sentry-sdk==1.3.1